Sound Manager Implementation

+++ gameMainWindow/mainwindow.cpp
@@ -50,7 +50,7 @@ void MainWindow::createActions(){
 
     setting = new Settings(this);
     connect(setting, SIGNAL(bgmAdjust(int)), this, SLOT(bgmAdjust(int)));
-    connect(setting, SIGNAL(sfxAdjust(int)), this, SLOT(sfxAdjust(int)));
+    connect(setting, SIGNAL(sfxAdjust(qreal)), this, SLOT(sfxAdjust(qreal)));
     connect(setting, SIGNAL(fullScreen(bool)), this, SLOT(fullScreen(bool)));
     setting->readSettings();
 }
